create finction checksDurationMeeting
